avoid size_t warnings
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@4745 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/goto-cc/run.cpp
@@ -63,7 +63,7 @@ int run(
 
   std::wstring wide_what=widen(what);
 
-  int status=_wspawnvp(_P_WAIT, wide_what.c_str(), _argv);
+  int status=(int)_wspawnvp(_P_WAIT, wide_what.c_str(), _argv);
 
   delete[] _argv;
